requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.1.1
schedule==1.2.1
pycoingecko==3.1.0 